% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownladMetaMeteoDataFrance.R
\name{DownloadFiles}
\alias{DownloadFiles}
\title{Donlowad MetaInfo and MetaData (for MeteoFrance Snow Open Data)}
\usage{

  DownloadFiles(url = c("https://donneespubliques.meteofrance.fr/client/document/doc_parametres_nivo_161.pdf",
  "https://donneespubliques.meteofrance.fr/donnees_libres/Txt/Nivo/postesNivo.csv",
  "https://donneespubliques.meteofrance.fr/donnees_libres/Txt/Nivo/postesNivo.json",
  "http://wiki.data.gouv.fr/images/9/9d/Licence_Ouverte.pdf"),
  dest = system.file("dest", package = "MeteoDataFrance"),
  method = c("wget", "curl"), force.download = FALSE, ...)
}
\arguments{
\item{url}{URLs. See the linked files in \url{https://donneespubliques.meteofrance.fr/?fond=produit&id_produit=94&id_rubrique=32}}

\item{dest}{destination directory where the downladed csv file is saved}

\item{method}{method applied in \code{\link{download.file}}}

\item{force.download}{logical value. If \code{TRUE} files are forced to be downloaded and then overwritten, if it is \code{FALSE} (default) and the files already exist in \code{destdir} directory are not downloaded.}

\item{...}{further argument for \code{download.file}}
}
\description{
Donlowad MetaInfo and MetaData (for MeteoFrance Snow Open Data)
}
\examples{

files <- DownloadFiles()
DownloadFiles(destdir='/home/ecor/Dropbox/R-packages/MeteoDataFrance/inst/dest')



}

